AP-15434: Make sizeof return long
AP-15434 (Adjust ChunkSize by analyzing memory-footprint)

+++ org.knime.core.columnar.testing/src/main/java/org/knime/core/columnar/testing/TestDoubleData.java
@@ -96,7 +96,7 @@ public final class TestDoubleData extends AbstractTestData implements DoubleWrit
     }
 
     @Override
-    public int sizeOf() {
+    public long sizeOf() {
         return m_numValues;
     }
