+++ org.knime.core.columnar/src/main/java/org/knime/core/columnar/batch/DefaultWriteBatch.java
@@ -103,8 +103,8 @@ public class DefaultWriteBatch implements WriteBatch {
     }
 
     @Override
-    public int sizeOf() {
-        return Arrays.stream(m_data).mapToInt(ReferencedData::sizeOf).sum();
+    public long sizeOf() {
+        return Arrays.stream(m_data).mapToLong(ReferencedData::sizeOf).sum();
     }
 
     @Override
